chore: split plugin files
