+++ internal/swarm/plugin_manager_test.go
